+++ tests/unit/test_analyzer.py
@@ -1,61 +1,70 @@
-# tests/unit/test_data_analyzer.py
-from unittest.mock import patch
+# tests/unit/test_analyzer.py
 
 import pandas as pd
 import pytest
 
-from food_analysis.core.analyzer import DataAnalyzer
+from food_analysis.core.analyzer import compute_recipe_stats, recipe_reviews
+
+# ---------------------------
+# Fixtures pour les données
+# ---------------------------
 
 
 @pytest.fixture
-def sample_interactions() -> pd.DataFrame:
-    """Crée un petit DataFrame de test pour les interactions."""
+def sample_recipes():
+    return pd.DataFrame({"id": [1, 2, 3], "name": ["Pasta", "Pizza", "Salad"]})
+
+
+@pytest.fixture
+def sample_interactions():
     return pd.DataFrame(
         {
-            "recipe_id": [1, 1, 1, 2, 2, 3],
-            "rating": [5, 4, 3, 5, 0, 2],
+            "user_id": [1, 2, 3, 4, 5],
+            "recipe_id": [1, 1, 2, 3, 3],
+            "rating": [5, 4, 3, 2, 1],
+            "date": [
+                "2023-01-01",
+                "2023-01-02",
+                "2023-01-03",
+                "2023-01-04",
+                "2023-01-05",
+            ],
+            "review": ["Good", "Ok", "Nice", "Bad", "Meh"],
         }
     )
 
 
-def test_plot_rating_distribution_calls_plt(sample_interactions):
-    analyzer = DataAnalyzer()
-
-    with (
-        patch("food_analysis.core.analyzer.plt.hist") as mock_hist,
-        patch("food_analysis.core.analyzer.plt.show") as mock_show,
-        patch("food_analysis.core.analyzer.plt.title") as mock_title,
-        patch("food_analysis.core.analyzer.plt.xlabel") as mock_xlabel,
-        patch("food_analysis.core.analyzer.plt.ylabel") as mock_ylabel,
-        patch("food_analysis.core.analyzer.plt.xlim") as mock_xlim,
-    ):
-        analyzer.plot_rating_distribution(sample_interactions, recipe_id=1)
-
-        # Convertir la Series en liste pour comparer
-        expected_ratings = sample_interactions.loc[
-            sample_interactions["recipe_id"] == 1, "rating"
-        ].tolist()
-
-        args, kwargs = mock_hist.call_args
-        assert list(args[0]) == expected_ratings
-        assert kwargs["bins"] == 6
-        assert kwargs["edgecolor"] == "black"
-
-        mock_title.assert_called_once_with("Distribution des notes pour la recette 1")
-        mock_xlabel.assert_called_once_with("Note")
-        mock_ylabel.assert_called_once_with("Nombre d'avis")
-        mock_xlim.assert_called_once_with(-0.25, None)
-        mock_show.assert_called_once()
+# ---------------------------
+# Tests compute_recipe_stats
+# ---------------------------
 
 
-def test_plot_rating_distribution_no_ratings() -> None:
-    """Teste le cas où la recette n'a pas de notes."""
-    analyzer = DataAnalyzer()
-    empty_df = pd.DataFrame({"recipe_id": [], "rating": []})
+def test_compute_recipe_stats(sample_recipes, sample_interactions):
+    result = compute_recipe_stats(sample_recipes, sample_interactions, m=1)
 
-    with patch("food_analysis.core.analyzer.plt") as mock_plt:
-        analyzer.plot_rating_distribution(empty_df, recipe_id=42)
+    # Vérifie les colonnes attendues
+    assert set(result.columns) == {"name", "avg_rating", "n_reviews", "weighted_rating"}
 
-        # Même sans données, plt.hist et plt.show doivent être appelés
-        mock_plt.hist.assert_called_once()
-        mock_plt.show.assert_called_once()
+    # Vérifie que le nombre de lignes correspond aux recettes avec interactions
+    assert len(result) == 3
+
+    # Vérifie que la note pondérée est calculée
+    assert all(result["weighted_rating"] > 0)
+
+
+# ---------------------------
+# Tests recipe_reviews
+# ---------------------------
+
+
+def test_recipe_reviews(sample_interactions):
+    df_reviews = recipe_reviews(3, sample_interactions)
+
+    # Vérifie que l'ID de recette correspond
+    assert all(df_reviews["user_id"].isin([4, 5]))
+
+    # Vérifie que les colonnes sont correctes
+    assert list(df_reviews.columns) == ["user_id", "rating", "date", "review"]
+
+    # Vérifie que le tri par date est correct (descendant)
+    assert df_reviews.iloc[0]["date"] >= df_reviews.iloc[1]["date"]
